+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 <!-- [UNRELEASED] -->
+## [1.0.1] 2023-07-25
+### Fix
+- Fix MySQL Database Service lifecycle event
+- Update tyapi-core library to 1.2.3
+
 ## [1.0.0] 2023-07-25
 ### Added
 - MySQL Database Service